controle de versionamento, ajuste no armazenamento de armas, carregadores e munições. integração com cfgGameplay, centralização de logs

+++ Scripts/3_Game/CfgGameplayHandler.c
@@ -27,7 +27,7 @@ modded class CfgGameplayHandler
 		return m_Data.VSMData.autoCloseInterval;
 	} 
 	//----------------------------------------------------------------------------------
-	static int GetVSM_AutoClosePlayerDistance()
+	static float GetVSM_AutoClosePlayerDistance()
 	{
 		return m_Data.VSMData.autoClosePlayerDistance;
 	}
